Moved top-level tests into core.
